Feature/order

+++ src/main/java/com/example/clothsdanawa/order/dto/OrderRequestDto.java
@@ -10,6 +10,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class OrderRequestDto {
 
-	private final Long point;
+	private final int point;
 
 }
